+++ packages/frontend/src/pages/page-editor/els/page-editor.el.note.vue
@@ -5,7 +5,7 @@ SPDX-License-Identifier: AGPL-3.0-only
 
 <template>
 <!-- eslint-disable vue/no-mutating-props -->
-<XContainer :draggable="true" :dragStartCallback="dragStartCallback" @remove="() => emit('remove')">
+<XContainer :draggable="true" :dragStartCallback="dragStartCallback" :touchStartCallback="touchStartCallback" @remove="() => emit('remove')">
 	<template #header><i class="ti ti-note"></i> {{ i18n.ts._pages.blocks.note }}</template>
 
 	<section style="padding: 16px;" class="_gaps_s">
@@ -35,6 +35,7 @@ import { i18n } from '@/i18n.js';
 
 const props = defineProps<{
 	dragStartCallback?: (ev: DragEvent) => void;
+	touchStartCallback?: (ev: TouchEvent) => void;
 	modelValue: Misskey.entities.PageBlock & { type: 'note' };
 }>();